+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_client.py
@@ -52,9 +52,6 @@ class ServiceBusClient(object):
     :keyword float retry_backoff_factor: Delta back-off internal in the unit of second between retries.
      Default value is 0.8.
     :keyword float retry_backoff_max: Maximum back-off interval in the unit of second. Default value is 120.
-    :keyword float timeout: Timeout setting for the operation in the unit of second, Default value is 60.
-     This timeout doesn't apply to `ServiceBusReceiver.receive` method and the iterator approach on ServiceBusReceiver
-     as they have dedicated parameters for timeout control.
 
     .. admonition:: Example:
 
@@ -148,9 +145,6 @@ class ServiceBusClient(object):
         :keyword float retry_backoff_factor: Delta back-off internal in the unit of second between retries.
          Default value is 0.8.
         :keyword float retry_backoff_max: Maximum back-off interval in the unit of second. Default value is 120.
-        :keyword float timeout: Timeout setting for the operation in the unit of second, Default value is 60.
-         This timeout doesn't apply to `ServiceBusReceiver.receive` method and the iterator approach on ServiceBusReceiver
-         as they have dedicated parameters for timeout control.
         :rtype: ~azure.servicebus.ServiceBusClient
 
         .. admonition:: Example:
@@ -205,7 +199,6 @@ class ServiceBusClient(object):
             retry_total=self._config.retry_total,
             retry_backoff_factor=self._config.retry_backoff_factor,
             retry_backoff_max=self._config.retry_backoff_max,
-            timeout=self._config.timeout,
             **kwargs
         )
         self._handlers.append(handler)
@@ -266,7 +259,6 @@ class ServiceBusClient(object):
             retry_total=self._config.retry_total,
             retry_backoff_factor=self._config.retry_backoff_factor,
             retry_backoff_max=self._config.retry_backoff_max,
-            timeout=self._config.timeout,
             **kwargs
         )
         self._handlers.append(handler)
@@ -301,7 +293,6 @@ class ServiceBusClient(object):
             retry_total=self._config.retry_total,
             retry_backoff_factor=self._config.retry_backoff_factor,
             retry_backoff_max=self._config.retry_backoff_max,
-            timeout=self._config.timeout,
             **kwargs
         )
         self._handlers.append(handler)
@@ -365,7 +356,6 @@ class ServiceBusClient(object):
                 retry_total=self._config.retry_total,
                 retry_backoff_factor=self._config.retry_backoff_factor,
                 retry_backoff_max=self._config.retry_backoff_max,
-                timeout=self._config.timeout,
                 **kwargs
             )
         else:
@@ -382,7 +372,6 @@ class ServiceBusClient(object):
                 retry_total=self._config.retry_total,
                 retry_backoff_factor=self._config.retry_backoff_factor,
                 retry_backoff_max=self._config.retry_backoff_max,
-                timeout=self._config.timeout,
                 **kwargs
             )
         self._handlers.append(handler)
@@ -441,7 +430,6 @@ class ServiceBusClient(object):
             retry_total=self._config.retry_total,
             retry_backoff_factor=self._config.retry_backoff_factor,
             retry_backoff_max=self._config.retry_backoff_max,
-            timeout=self._config.timeout,
             **kwargs
         )
         self._handlers.append(handler)
@@ -498,7 +486,6 @@ class ServiceBusClient(object):
             retry_total=self._config.retry_total,
             retry_backoff_factor=self._config.retry_backoff_factor,
             retry_backoff_max=self._config.retry_backoff_max,
-            timeout=self._config.timeout,
             **kwargs
         )
         self._handlers.append(handler)